+++ engine/group.h
@@ -4,5 +4,5 @@
 #include "node.h"
 
 void *perform_ensemble(Node *node, ensemble_state *state, Node *_inputs[],
-                       int nframes, double spf);
+                       int nframes, float spf);
 #endif // GROUP_H
